refactor(devtools): remove unneeded tsconfig keyword args for ts_project
These keyword args are not needed because `ts_project` and `ng_ts_project` already default to the tsconfig at the root of the devtools directory.

This commit removes these key word args from places where tsconfig was used as a keyword arg with the root level tsconfig as the target.