feat(devtools): create profiler for injector events in dev mode and supported debug APIs
Motivation:
Currently, understanding dependency injection in Angular requires a lot of context and has been sited in our surveys as one of the largest points of confusion that our users have with the framework. This commit is the beginning of our approach to make debugging dependency injection in Angular easier.

Overview:
This commit introduces injector profiler callbacks in parts of the framework to emit injector events. It also introduces a default handler for these events. This default handler parses the stream of events to construct some data structures that will support new injector debug APIs.

This commit introduces 3 new APIs for each of the following use cases
- Given an injector and a token that was instantiated by that injector, discover all of the things were injected in the instance of that token
- Given an injector, discover all of the providers that were configured in that injector.
- Given an injector, discover the parent of that injector.

Why was this implemented with a profiler?:
We have implemented a similar pattern in the past to minimize overhead. There is also the possiblity of making the internal `setInjectorProfiler` function a public debug API in the future, so that users can implement their own handlers to debug DI events. Lastly DI in Angular maps nicely to a stream of events.

Will this affect runtime performance?:
For production applications no. For applications in dev mode there is some additional overhead from the default profiler handling injector events and holding debug data in memory.

Will this affect bundle size?:
No. Dead code elimination should strip all the code used by this commit.